HW2 partial completed without filtering on meal and user

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepository.java
@@ -1,5 +1,6 @@
 package ru.javawebinar.topjava.repository.inmemory;
 
+import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
@@ -8,19 +9,25 @@ import java.util.Collection;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
 
+@Repository
 public class InMemoryMealRepository implements MealRepository {
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
     private AtomicInteger counter = new AtomicInteger(0);
 
     {
-        MealsUtil.MEALS.forEach(this::save);
+//        MealsUtil.MEALS.forEach(meal -> save(meal, 1));
+        for (Meal meal: MealsUtil.MEALS) {
+            save(meal, 1);
+        }
     }
 
     @Override
-    public Meal save(Meal meal) {
+    public Meal save(Meal meal, int authUserId) {
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
+            meal.setUserId(authUserId);
             repository.put(meal.getId(), meal);
             return meal;
         }
@@ -29,18 +36,41 @@ public class InMemoryMealRepository implements MealRepository {
     }
 
     @Override
-    public boolean delete(int id) {
+    public boolean delete(int id, int authUserId) {
+        if (wrongUserId(id, authUserId)) {
+            return false;
+        }
         return repository.remove(id) != null;
     }
 
     @Override
-    public Meal get(int id) {
+    public Meal get(int id, int authUserId) {
+        if (wrongUserId(id, authUserId)) {
+            return null;
+        }
         return repository.get(id);
     }
 
     @Override
-    public Collection<Meal> getAll() {
-        return repository.values();
+    public Collection<Meal> getAll(int authUserId) {
+        return repository.values().stream()
+                .filter(meal -> {
+                    System.out.println(meal.getUserId() + " "+ authUserId);
+                    return meal.getUserId().equals(authUserId);
+                })
+                .collect(Collectors.toList());
+    }
+
+    private boolean wrongUserId(int mealId, int authUserId) {
+        if (repository.get(mealId) == null) {
+            return true;
+        }
+        return !repository.get(mealId).getUserId().equals(authUserId);
+    }
+
+    @Override
+    public Collection<Meal> filterByDate(int authUserId) {
+        return getAll(authUserId);
     }
 }
